LayoutTests/platform/chromium/inspector/console/console-uncaught-exception-expected.txt

CONSOLE MESSAGE: line 18: Uncaught Error: Exception in inline script.
CONSOLE MESSAGE: line 11: Uncaught Error: Exception in 'load' event listener.
CONSOLE MESSAGE: line 6: Uncaught Error: Exception in setTimeout callback.
Tests that uncaught exceptions are logged into console.Bug 47250.


Uncaught Error: Exception in inline script. uncaught-in-iframe.html:18a uncaught-in-iframe.html:18b uncaught-in-iframe.html:23(anonymous function) uncaught-in-iframe.html:26
Uncaught Error: Exception in 'load' event listener. uncaught-in-iframe.html:11f uncaught-in-iframe.html:11(anonymous function) uncaught-in-iframe.html:13
Uncaught Error: Exception in setTimeout callback. uncaught-in-iframe.html:6bar uncaught-in-iframe.html:6(anonymous function) uncaught-in-iframe.html:8
